better N1QL query for root data & fix deep child added unhandled rejection
